+++ custom_components/thz/cop_sensor.py
@@ -148,7 +148,7 @@ async def async_setup_cop_sensors(
 
     if cop_sensors:
         assign_subdevices(cop_sensors, config_entry.data)
-        async_add_entities(cop_sensors, True)
+        async_add_entities(cop_sensors)
         _LOGGER.debug("Created %d COP sensors", len(cop_sensors))
     else:
         _LOGGER.debug("No COP sensors could be created - missing required data")